# internal/transport/http/swaggerdocs/swagger.yaml
basePath: /api/v1
definitions:
...
  shared.ErrorBody:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  shared.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/shared.ErrorBody'
    type: object
...
  shared.HonorGroupResponse:
    properties:
      backgroundAssetbundleName:
        type: string
      frameName:
        type: string
      honorType:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  shared.HonorLevelResponse:
    properties:
      assetbundleName:
        type: string
      bonus:
        type: number
      description:
        type: string
      honorId:
        type: integer
      honorRarity:
        type: string
      level:
        type: integer
    type: object
  shared.HonorListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/shared.HonorObjectResponse'
        type: array
      pagination:
        $ref: '#/definitions/shared.PaginationResponse'
    type: object
  shared.HonorObjectResponse:
    properties:
      assetbundleName:
        type: string
      group:
        $ref: '#/definitions/shared.HonorGroupResponse'
      groupId:
        type: integer
      honorMissionType:
        type: string
      honorRarity:
        type: string
      honorType:
        type: string
      id:
        type: integer
      levels:
        items:
          $ref: '#/definitions/shared.HonorLevelResponse'
        type: array
      name:
        type: string
      seq:
        type: integer
    type: object
...
  shared.PaginationResponse:
    properties:
      has_next:
        type: boolean
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
info:
  contact: {}
  description: API for master data sync and card querying.
  title: sekai-master-api
  version: "1.0"
paths:
...
  /honors/{region}/{id}:
    get:
      parameters:
      - description: Region
        in: path
        name: region
        required: true
        type: string
      - description: Honor ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/shared.HonorObjectResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      summary: Get an honor by id
      tags:
      - honors
  /honors/{region}/list:
    get:
      parameters:
      - description: Region
        in: path
        name: region
        required: true
        type: string
      - description: Page number
        in: query
        minimum: 1
        name: page
        type: integer
      - description: Page size
        in: query
        maximum: 100
        minimum: 1
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/shared.HonorListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      summary: List honors by page
      tags:
      - honors
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"